+++ packages/rs-drive-abci/src/execution/platform_events/fee_pool_outwards_distribution/add_distribute_fees_from_oldest_unpaid_epoch_pool_to_proposers_operations/v0/mod.rs
@@ -92,7 +92,7 @@ mod tests {
     use super::*;
     use dpp::block::block_info::BlockInfo;
 
-    use drive::common::identities::create_test_masternode_identities_and_add_them_as_epoch_block_proposers;
+    use drive::common::test_utils::identities::create_test_masternode_identities_and_add_them_as_epoch_block_proposers;
 
     use crate::test::helpers::setup::TestPlatformBuilder;
