+++ packages/rs-drive-abci/src/execution/platform_events/initialization/create_genesis_state/v0/mod.rs
@@ -314,8 +314,7 @@ mod tests {
             assert_eq!(
                 root_hash,
                 [
-                    144, 123, 124, 62, 37, 243, 19, 45, 226, 49, 97, 251, 141, 197, 94, 69, 117,
-                    126, 182, 44, 148, 143, 41, 113, 245, 86, 229, 207, 100, 56, 17, 177
+                    145, 134, 225, 227, 181, 246, 225, 73, 216, 216, 249, 73, 21, 213, 116, 110, 85, 106, 80, 222, 95, 167, 97, 242, 69, 142, 176, 211, 89, 182, 162, 67
                 ]
             )
         }
