+++ packages/rs-drive-abci/src/query/identity_based_queries/identity_contract_nonce/v0/mod.rs
@@ -85,7 +85,7 @@ mod tests {
     use dpp::identity::identity_nonce::{
         IDENTITY_NONCE_VALUE_FILTER, IDENTITY_NONCE_VALUE_FILTER_MAX_BYTES,
     };
-    use drive::common::identities::create_test_identity_with_rng;
+    use drive::common::test_utils::identities::create_test_identity_with_rng;
     use rand::rngs::StdRng;
     use rand::{Rng, SeedableRng};
