+++ packages/rs-drive-abci/src/test/helpers/fee_pools.rs
@@ -49,7 +49,7 @@ use dpp::document::{DocumentV0, INITIAL_REVISION};
 use dpp::identity::accessors::IdentityGettersV0;
 use dpp::system_data_contracts::masternode_reward_shares_contract::v1::document_types;
 use dpp::version::PlatformVersion;
-use drive::common::identities::create_test_identity_with_rng;
+use drive::common::test_utils::identities::create_test_identity_with_rng;
 use drive::dpp::document::Document;
 use drive::drive::flags::StorageFlags;
 use drive::drive::object_size_info::DocumentInfo::DocumentRefInfo;
