+++ packages/rs-drive/Cargo.toml
@@ -46,10 +46,10 @@ enum-map = { version = "2.0.3", optional = true }
 intmap = { version = "2.0.0", features = ["serde"], optional = true }
 chrono = { version = "0.4.35", optional = true }
 itertools = { version = "0.11.0", optional = true }
-grovedb = { git = "https://github.com/dashpay/grovedb", rev = "9eae5457f18d307d673767c8d772630a9d0ad66d", optional = true, default-features = false }
-grovedb-costs = { git = "https://github.com/dashpay/grovedb", rev = "9eae5457f18d307d673767c8d772630a9d0ad66d", optional = true }
-grovedb-path = { git = "https://github.com/dashpay/grovedb", rev = "9eae5457f18d307d673767c8d772630a9d0ad66d" }
-grovedb-storage = { git = "https://github.com/dashpay/grovedb", rev = "9eae5457f18d307d673767c8d772630a9d0ad66d", optional = true }
+grovedb = { git = "https://github.com/dashpay/grovedb", rev = "f5a3382c5537de7417e983cdf2132792aa9fdc9d", optional = true, default-features = false }
+grovedb-costs = { git = "https://github.com/dashpay/grovedb", rev = "f5a3382c5537de7417e983cdf2132792aa9fdc9d", optional = true }
+grovedb-path = { git = "https://github.com/dashpay/grovedb", rev = "f5a3382c5537de7417e983cdf2132792aa9fdc9d" }
+grovedb-storage = { git = "https://github.com/dashpay/grovedb", rev = "f5a3382c5537de7417e983cdf2132792aa9fdc9d", optional = true }
 
 [dev-dependencies]
 criterion = "0.3.5"
