+++ packages/rs-drive/src/drive/grove_operations/batch_insert_if_not_exists/v0/mod.rs
@@ -94,7 +94,7 @@ impl Drive {
                             GroveDb::average_case_for_has_raw(
                                 &key_info_path,
                                 &key_info,
-                                element.serialized_size() as u32,
+                                element.serialized_size()? as u32,
                                 in_tree_using_sums,
                             ),
                         ));
