+++ packages/rs-drive/tests/deterministic_root_hash.rs
@@ -446,7 +446,7 @@ fn test_root_hash_with_batches(drive: &Drive, db_transaction: &Transaction) {
         .unwrap()
         .expect("should return app hash");
 
-    let expected_app_hash = "6b8bbf1f069858abf57573f43a62e27d60e6139c4d23e1fe572fa3fe34057973";
+    let expected_app_hash = "8880c628746348b09621b00589ccf856f75634388436a105473eb4506bb7e85e";
 
     assert_eq!(hex::encode(app_hash), expected_app_hash);
 }
