+++ packages/rs-drive/tests/query_tests.rs
@@ -993,8 +993,7 @@ fn test_family_basic_queries() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        66, 138, 16, 202, 64, 147, 26, 115, 247, 29, 235, 148, 80, 154, 249, 204, 148, 59, 248, 83,
-        118, 140, 196, 38, 24, 59, 190, 48, 212, 130, 220, 215,
+163, 56, 30, 144, 201, 147, 186, 46, 12, 101, 223, 105, 199, 6, 242, 76, 137, 232, 95, 87, 56, 167, 65, 114, 254, 221, 132, 182, 221, 89, 197, 125,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -2305,8 +2304,7 @@ fn test_family_basic_queries() {
     assert_eq!(
         root_hash.as_slice(),
         vec![
-            183, 156, 84, 144, 94, 45, 111, 118, 255, 200, 166, 248, 252, 127, 216, 105, 83, 124,
-            126, 241, 19, 191, 32, 188, 112, 101, 249, 194, 188, 224, 188, 162
+            173, 209, 205, 148, 17, 96, 43, 47, 195, 124, 81, 242, 237, 221, 130, 189, 203, 158, 19, 6, 217, 19, 231, 186, 70, 66, 1, 105, 7, 78, 245, 63,
         ],
     );
 }
@@ -2452,8 +2450,7 @@ fn test_family_starts_at_queries() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        66, 138, 16, 202, 64, 147, 26, 115, 247, 29, 235, 148, 80, 154, 249, 204, 148, 59, 248, 83,
-        118, 140, 196, 38, 24, 59, 190, 48, 212, 130, 220, 215,
+163, 56, 30, 144, 201, 147, 186, 46, 12, 101, 223, 105, 199, 6, 242, 76, 137, 232, 95, 87, 56, 167, 65, 114, 254, 221, 132, 182, 221, 89, 197, 125,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -2894,8 +2891,7 @@ fn test_family_with_nulls_query() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        120, 51, 207, 201, 227, 63, 44, 130, 175, 117, 220, 200, 56, 145, 137, 7, 203, 88, 178, 88,
-        211, 192, 197, 201, 253, 79, 18, 50, 43, 156, 204, 153,
+        181, 187, 251, 87, 13, 132, 222, 134, 83, 231, 5, 226, 79, 212, 220, 50, 123, 224, 81, 128, 27, 149, 15, 194, 46, 48, 208, 157, 77, 175, 92, 246,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -3017,8 +3013,7 @@ fn test_query_with_cached_contract() {
 
     // Make sure the state is deterministic
     let expected_app_hash = vec![
-        66, 138, 16, 202, 64, 147, 26, 115, 247, 29, 235, 148, 80, 154, 249, 204, 148, 59, 248, 83,
-        118, 140, 196, 38, 24, 59, 190, 48, 212, 130, 220, 215,
+163, 56, 30, 144, 201, 147, 186, 46, 12, 101, 223, 105, 199, 6, 242, 76, 137, 232, 95, 87, 56, 167, 65, 114, 254, 221, 132, 182, 221, 89, 197, 125,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -3166,8 +3161,7 @@ fn test_dpns_query() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        104, 127, 253, 216, 170, 92, 9, 250, 226, 44, 134, 38, 89, 10, 135, 4, 121, 65, 255, 136,
-        149, 217, 217, 169, 108, 37, 241, 86, 57, 214, 131, 96,
+88, 19, 180, 210, 138, 212, 101, 137, 172, 158, 42, 123, 153, 162, 58, 30, 36, 3, 55, 17, 211, 254, 39, 224, 94, 181, 43, 18, 198, 24, 197, 20,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -3718,8 +3712,7 @@ fn test_dpns_query_start_at() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        104, 127, 253, 216, 170, 92, 9, 250, 226, 44, 134, 38, 89, 10, 135, 4, 121, 65, 255, 136,
-        149, 217, 217, 169, 108, 37, 241, 86, 57, 214, 131, 96,
+88, 19, 180, 210, 138, 212, 101, 137, 172, 158, 42, 123, 153, 162, 58, 30, 36, 3, 55, 17, 211, 254, 39, 224, 94, 181, 43, 18, 198, 24, 197, 20,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash,);
@@ -3812,8 +3805,7 @@ fn test_dpns_query_start_after() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        104, 127, 253, 216, 170, 92, 9, 250, 226, 44, 134, 38, 89, 10, 135, 4, 121, 65, 255, 136,
-        149, 217, 217, 169, 108, 37, 241, 86, 57, 214, 131, 96,
+88, 19, 180, 210, 138, 212, 101, 137, 172, 158, 42, 123, 153, 162, 58, 30, 36, 3, 55, 17, 211, 254, 39, 224, 94, 181, 43, 18, 198, 24, 197, 20,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -3906,8 +3898,7 @@ fn test_dpns_query_start_at_desc() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        104, 127, 253, 216, 170, 92, 9, 250, 226, 44, 134, 38, 89, 10, 135, 4, 121, 65, 255, 136,
-        149, 217, 217, 169, 108, 37, 241, 86, 57, 214, 131, 96,
+88, 19, 180, 210, 138, 212, 101, 137, 172, 158, 42, 123, 153, 162, 58, 30, 36, 3, 55, 17, 211, 254, 39, 224, 94, 181, 43, 18, 198, 24, 197, 20,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -4000,8 +3991,7 @@ fn test_dpns_query_start_after_desc() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        104, 127, 253, 216, 170, 92, 9, 250, 226, 44, 134, 38, 89, 10, 135, 4, 121, 65, 255, 136,
-        149, 217, 217, 169, 108, 37, 241, 86, 57, 214, 131, 96,
+88, 19, 180, 210, 138, 212, 101, 137, 172, 158, 42, 123, 153, 162, 58, 30, 36, 3, 55, 17, 211, 254, 39, 224, 94, 181, 43, 18, 198, 24, 197, 20,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -4192,8 +4182,7 @@ fn test_dpns_query_start_at_with_null_id() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        11, 19, 155, 212, 245, 198, 124, 117, 180, 86, 195, 68, 236, 38, 209, 25, 16, 186, 122, 12,
-        9, 73, 22, 156, 179, 16, 190, 94, 19, 60, 210, 82,
+16, 77, 89, 81, 202, 42, 156, 156, 114, 212, 3, 107, 144, 120, 108, 241, 67, 12, 234, 138, 245, 81, 48, 252, 72, 87, 16, 96, 118, 181, 12, 169,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -4393,8 +4382,7 @@ fn test_dpns_query_start_after_with_null_id() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        11, 19, 155, 212, 245, 198, 124, 117, 180, 86, 195, 68, 236, 38, 209, 25, 16, 186, 122, 12,
-        9, 73, 22, 156, 179, 16, 190, 94, 19, 60, 210, 82,
+16, 77, 89, 81, 202, 42, 156, 156, 114, 212, 3, 107, 144, 120, 108, 241, 67, 12, 234, 138, 245, 81, 48, 252, 72, 87, 16, 96, 118, 181, 12, 169,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash);
@@ -4598,8 +4586,7 @@ fn test_dpns_query_start_after_with_null_id_desc() {
         .expect("there is always a root hash");
 
     let expected_app_hash = vec![
-        11, 19, 155, 212, 245, 198, 124, 117, 180, 86, 195, 68, 236, 38, 209, 25, 16, 186, 122, 12,
-        9, 73, 22, 156, 179, 16, 190, 94, 19, 60, 210, 82,
+16, 77, 89, 81, 202, 42, 156, 156, 114, 212, 3, 107, 144, 120, 108, 241, 67, 12, 234, 138, 245, 81, 48, 252, 72, 87, 16, 96, 118, 181, 12, 169,
     ];
 
     assert_eq!(root_hash.as_slice(), expected_app_hash,);
