+++ packages/rs-drive/tests/query_tests_history.rs
@@ -291,8 +291,7 @@ fn test_query_historical() {
     assert_eq!(
         root_hash.as_slice(),
         vec![
-            183, 18, 172, 146, 135, 191, 55, 236, 73, 61, 68, 217, 55, 134, 87, 148, 210, 99, 196,
-            151, 94, 207, 218, 133, 70, 221, 228, 57, 197, 93, 12, 45,
+            181, 185, 108, 50, 217, 141, 128, 232, 165, 226, 95, 0, 177, 127, 230, 52, 238, 131, 107, 59, 237, 63, 184, 244, 2, 66, 114, 195, 97, 175, 252, 209,
         ]
     );
 
@@ -1655,8 +1654,7 @@ fn test_query_historical() {
     assert_eq!(
         root_hash.as_slice(),
         vec![
-            225, 224, 171, 0, 235, 104, 240, 209, 248, 74, 159, 9, 246, 185, 174, 114, 53, 133,
-            235, 206, 36, 151, 226, 124, 108, 205, 152, 122, 6, 202, 177, 210
+            25, 175, 56, 55, 43, 121, 224, 243, 179, 106, 24, 41, 94, 74, 167, 144, 67, 125, 213, 119, 247, 9, 237, 191, 50, 90, 22, 115, 131, 152, 246, 102
         ]
     );
 }
